refactor(web): move mappers to feat dirs
refactor(web): fix gammy paths remove old code
