+++ appeals/web/src/server/appeals/appeal-details/appellant-case/appellant-case.mapper.js
@@ -26,9 +26,9 @@ import {
 import { DEADLINE_HOUR, DEADLINE_MINUTE } from '@pins/appeals/constants/dates.js';
 import { isFeatureActive } from '#common/feature-flags.js';
 import { APPEAL_TYPE, FEATURE_FLAG_NAMES } from '@pins/appeals/constants/common.js';
-import { generateHASComponents } from '#lib/mappers/appellant-case/appeal-type-has.mapper.js';
-import { generateFPAComponents } from '#lib/mappers/appellant-case/appeal-type-fpa.mapper.js';
 import { userHasPermission } from '#lib/mappers/index.js';
+import { generateHASComponents } from './page-components/has.mapper.js';
+import { generateFPAComponents } from './page-components/fpa.mapper.js';
 import { permissionNames } from '#environment/permissions.js';
 
 /**
