+++ appeals/web/src/server/appeals/appeal-details/appellant-case/page-components/has.mapper.js
