+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/view-and-review/page-components/common.js
@@ -4,7 +4,7 @@ import { simpleHtmlComponent, wrapComponents } from '#lib/mappers/components/htm
 import { buildHtmUnorderedList } from '#lib/nunjucks-template-builders/tag-builders.js';
 import { COMMENT_STATUS } from '@pins/appeals/constants/common.js';
 
-/** @typedef {import("../../../../../appeals/appeal-details/interested-party-comments/interested-party-comments.types.js").Representation} Representation */
+/** @typedef {import("../../interested-party-comments.types.js").Representation} Representation */
 
 /**
  * Generates the withdraw link component.
