+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/view-and-review/page-components/view.mapper.js
@@ -2,8 +2,8 @@ import { appealShortReference } from '#lib/appeals-formatter.js';
 import { buildNotificationBanners } from '#lib/mappers/notification-banners.mapper.js';
 import { generateCommentSummaryList, generateWithdrawLink } from './common.js';
 
-/** @typedef {import("../../../../../appeals/appeal-details/appeal-details.types.js").WebAppeal} Appeal */
-/** @typedef {import("../../../../../appeals/appeal-details/interested-party-comments/interested-party-comments.types.js").Representation} Representation */
+/** @typedef {import("#appeals/appeal-details/appeal-details.types.js").WebAppeal} Appeal */
+/** @typedef {import("#appeals/appeal-details/interested-party-comments/interested-party-comments.types.js").Representation} Representation */
 
 /**
  * @param {Appeal} appealDetails
